backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
torch
torchvision
torchxrayvision
grad-cam
pillow
numpy
matplotlib
google-genai